rename to WeatherProvider. move fetching/state into provider
